requirements.txt
selenium>=4.15.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pyperclip>=1.8.2
google-generativeai>=0.3.0
pyinstaller>=6.0.0
tkcalendar>=1.6.1
